ECOM- Adding method addProduct to add the product with a categoryID

+++ src/main/java/com/ecommerce/sb_ecom_project/payload/CategoryDTO.java
@@ -11,9 +11,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CategoryDTO {
-    private long categoryId;
+    private Integer categoryId;
     private String categoryName;
     private String categoryDescription;
     private boolean markActive;
-
 }
